New logo color WIP
The old logo doesn't look good with the fresh color scheme so some
tweaking was needed. We're using a PNG here, so it's not as easy as
it might be with a pure SVG.

I'm using Inkscape on the ec-logo-final.svg to tweak the colors, and
then exporting a PNG file.

WIP, since I'm not happy with this color.