Check in data flow diagram svg and pdf
This diagram is useful for our annual SOAR. I extracted it from a
LucidChart doc[1] which was embedded in a threat model doc from
2022[2], then updated the name with Inkscape[3].

I did all that so I could resubmit it on SOAR-3633.

Since we'll likely need this again next year, I decided to git add
it here so I we can find it again.

(It might also be useful as a starting point, or inspiration for,
diagrams that we include in our documenation. See EC-1270.)

Ref: https://issues.redhat.com/browse/EC-1430

[1] https://lucid.app/lucidchart/b41d49c1-0bd8-4410-9756-5fafa8817922/edit?page=0_0#
[2] https://docs.google.com/document/d/1DNeEJQU6CUW1AVQp1z0H3BSl-mGc8-hyZNUljT8xkBA/edit?tab=t.0#heading=h.clj3egttwf1z
[3] https://inkscape.org/
[4] https://issues.redhat.com/browse/SOAR-3633
[5] https://issues.redhat.com/browse/EC-1270